requirements.txt
# langchain
streamlit
langchain
wikipedia
python-dotenv

# GPT file
openai
wikipedia-api

requests
tiktoken

# embedding 
faiss-cpu
pinecone-clien